topic/sensor-utilities: 	- filled in config values/structures 	- added dummy config values for fuel level and oil pressure 	- todo: figure out best way to hook up analog inputs to sensor configs

+++ QtDash/VolvoDigitalDashModels/app/mcp23017.h
@@ -161,7 +161,7 @@ public:
         int ret = i2c_smbus_read_byte_data(mFd, (uint8_t) reg);
 
         if (ret < 0) {
-            printf("i2c-%d device @0x%02X register addr: 0x%02X read failed: %d\n", mBus, mAddr, reg, errno);
+            printf("i2c-%d device @0x%02X register addr: 0x%02X read failed: %d\n", mBus, (unsigned int)mAddr, (unsigned int)reg, errno);
             return -1;
         }
